requirements.txt
loguru
PyYAML
requests
